# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.PaymentWebhookSample:
    properties:
      amount:
        example: 100.5
        type: number
      currency:
        example: USD
        type: string
      customer_id:
        example: cust_abc123
        type: string
      event_type:
        example: payment.completed
        type: string
      metadata:
        additionalProperties: true
        type: object
      payment_method:
        example: credit_card
        type: string
      status:
        example: completed
        type: string
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
      transaction_id:
        example: txn_123456789
        type: string
    type: object
...
  types.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  types.Feature:
    properties:
      description:
        type: string
      limited:
        type: boolean
      maxQuota:
        type: integer
      name:
        type: string
      recurringInterval:
        $ref: '#/definitions/types.RecurringInterval'
      slug:
        $ref: '#/definitions/types.FeatureSlug'
    type: object
  types.FeatureSlug:
    enum:
    - profile_optimizer
    - ai_proposal_credit
    - ai_proposals_unlimited
    - linkedin_profile_optimizer
    - priority_support
    - resume_generator
    - advanced_ai_insights
    - early_access_features
    - direct_support_chat
    - standard_support
    type: string
    x-enum-varnames:
    - FeatureProfileOptimizer
    - FeatureAIProposalCredit
    - FeatureAIProposalsUnlimited
    - FeatureLinkedInOptimizer
    - FeaturePrioritySupport
    - FeatureResumeGenerator
    - FeatureAdvancedAIInsights
    - FeatureEarlyAccess
    - FeatureDirectSupportChat
    - FeatureStandardSupport
...
  types.PlanRecurringInterval:
    enum:
    - monthly
    type: string
    x-enum-varnames:
    - PlanMonthly
  types.RecurringInterval:
    enum:
    - daily
    - weekly
    - monthly
    type: string
    x-enum-varnames:
    - Daily
    - Weekly
    - Monthly
  types.Tier:
    properties:
      createdAt:
        type: string
      description:
        type: string
      features:
        items:
          $ref: '#/definitions/types.Feature'
        type: array
      name:
        type: string
      polarRefId:
        type: string
      price:
        type: integer
      recurringInterval:
        $ref: '#/definitions/types.PlanRecurringInterval'
      slug:
        type: string
      updatedAt:
        type: string
    type: object
host: provolo-backend.onrender.com
info:
  contact: {}
  description: This is the Provolo backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Provolo API
  version: "1.0"
paths:
...
  /api/v1/payment/tiers:
    get:
      description: Retrieves all available payment tiers
      produces:
      - application/json
      responses:
        "200":
          description: Payment tiers retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Tier'
                  type: array
              type: object
        "500":
          description: Internal Server Error - Failed to retrieve payment tiers
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get payment tiers
      tags:
      - payments
  /api/v1/payment/tiers/{slug}:
    get:
      description: Retrieves a specific payment tier using its slug
      parameters:
      - description: Tier slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment tier retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.Tier'
              type: object
        "404":
          description: Tier not found
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get payment tier by slug
      tags:
      - payments
  /api/v1/payment/webhook:
    post:
      consumes:
      - application/json
      description: Handles payment webhook notifications from payment providers -
        accepts any JSON structure
      parameters:
      - description: Sample payment webhook structure (accepts any JSON)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PaymentWebhookSample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Handle payment webhook
      tags:
      - payments
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"